scripts/requirements.txt
tvscreener>=0.2.0
supabase>=2.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
